Tests/LibWeb/Text/expected/Editing/execCommand-insertOrderedList.txt

<ol><li>foobar</li></ol>
<div>foobar</div>
<ol><li>foobar</li></ol>
